Split DrawPoint to multiple files
