+++ lib/dashboard/ActionButtons.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:fluffy_board/dashboard/filemanager/AddFolder.dart';
 import 'package:fluffy_board/dashboard/filemanager/AddOfflineWhiteboard.dart';
 import 'package:fluffy_board/utils/ThemeDataUtils.dart';
-import 'package:fluffy_board/whiteboard/DrawPoint.dart';
+import 'package:fluffy_board/whiteboard/whiteboard-data/json_encodable.dart';
 import 'package:flutter/material.dart';
 import 'dart:ui' as ui;
 import 'package:localstorage/localstorage.dart';
