+++ lib/whiteboard/overlays/Toolbar/SettingsToolbar/TextItemSettings.dart
@@ -5,10 +5,11 @@ import 'package:fluffy_board/whiteboard/InfiniteCanvas.dart';
 import 'package:fluffy_board/whiteboard/Websocket/WebsocketConnection.dart';
 import 'package:fluffy_board/whiteboard/Websocket/WebsocketSend.dart';
 import 'package:fluffy_board/whiteboard/Websocket/WebsocketTypes.dart';
+import 'package:fluffy_board/whiteboard/whiteboard-data/textitem.dart';
 import 'package:flutter/material.dart';
 import 'dart:ui';
 
-import '../../../DrawPoint.dart';
+import '../../../whiteboard-data/json_encodable.dart';
 import '../../../WhiteboardView.dart';
 import '../../Toolbar.dart' as Toolbar;
 import 'package:sleek_circular_slider/sleek_circular_slider.dart';
