+++ lib/whiteboard/whiteboard-data/json_encodable.dart
@@ -1,0 +1,14 @@
+abstract class JsonEncodable {
+  toJSONEncodable();
+  JsonEncodable.fromJson(json);
+}
+
+
+
+
+
+
+
+
+
+
